requirements.txt
numpy
git+https://github.com/openai/CLIP.git
torch
torchvision
matplotlib
pillow  
scikit-image    
ipykernel
ftfy